Refactored day 5 parsing

+++ implementations/rust/src/years/y2022/day_5/part_1.rs
@@ -1,10 +1,10 @@
 use crate::years::y2022::day_5::common::{parse_input, take_tops};
 
-fn execute_commands(columns: &mut Vec<Vec<char>>, instructions: Vec<[i8; 3]>) {
+fn execute_commands(columns: &mut Vec<Vec<char>>, instructions: Vec<[usize; 3]>) {
     instructions.into_iter().for_each(|[n, from, to]| {
         for _ in 0..n {
-            let top = columns[from as usize].pop().expect("Top from column");
-            columns[to as usize].push(top);
+            let top = columns[from].pop().expect("Top from column");
+            columns[to].push(top);
         }
     })
 }
